Flytter Brevperiodegenerering inn i brevperiodegenerator
